bindings: generate uses entities/types
