+++ pkg/bindings/generate/generate.go
@@ -7,10 +7,10 @@ import (
 	"strconv"
 
 	"github.com/containers/podman/v4/pkg/bindings"
-	"github.com/containers/podman/v4/pkg/domain/entities"
+	"github.com/containers/podman/v4/pkg/domain/entities/types"
 )
 
-func Systemd(ctx context.Context, nameOrID string, options *SystemdOptions) (*entities.GenerateSystemdReport, error) {
+func Systemd(ctx context.Context, nameOrID string, options *SystemdOptions) (*types.GenerateSystemdReport, error) {
 	if options == nil {
 		options = new(SystemdOptions)
 	}
@@ -29,14 +29,14 @@ func Systemd(ctx context.Context, nameOrID string, options *SystemdOptions) (*en
 	}
 	defer response.Body.Close()
 
-	report := &entities.GenerateSystemdReport{}
+	report := &types.GenerateSystemdReport{}
 	return report, response.Process(&report.Units)
 }
 
 // Kube generate Kubernetes YAML (v1 specification)
 //
 // Note: Caller is responsible for closing returned reader
-func Kube(ctx context.Context, nameOrIDs []string, options *KubeOptions) (*entities.GenerateKubeReport, error) {
+func Kube(ctx context.Context, nameOrIDs []string, options *KubeOptions) (*types.GenerateKubeReport, error) {
 	if options == nil {
 		options = new(KubeOptions)
 	}
@@ -64,7 +64,7 @@ func Kube(ctx context.Context, nameOrIDs []string, options *KubeOptions) (*entit
 	}
 
 	if response.StatusCode == http.StatusOK {
-		return &entities.GenerateKubeReport{Reader: response.Body}, nil
+		return &types.GenerateKubeReport{Reader: response.Body}, nil
 	}
 
 	// Unpack the error.
