+++ pkg/domain/entities/generate.go
@@ -1,6 +1,8 @@
 package entities
 
-import "io"
+import (
+	"github.com/containers/podman/v4/pkg/domain/entities/types"
+)
 
 // GenerateSystemdOptions control the generation of systemd unit files.
 type GenerateSystemdOptions struct {
@@ -22,10 +24,7 @@ type GenerateSystemdOptions struct {
 }
 
 // GenerateSystemdReport
-type GenerateSystemdReport struct {
-	// Units of the generate process. key = unit name -> value = unit content
-	Units map[string]string
-}
+type GenerateSystemdReport = types.GenerateSystemdReport
 
 // GenerateKubeOptions control the generation of Kubernetes YAML files.
 type GenerateKubeOptions struct {
@@ -44,16 +43,9 @@ type GenerateKubeOptions struct {
 type KubeGenerateOptions = GenerateKubeOptions
 
 // GenerateKubeReport
-//
-// FIXME: Podman4.0 should change io.Reader to io.ReaderCloser
-type GenerateKubeReport struct {
-	// Reader - the io.Reader to reader the generated YAML file.
-	Reader io.Reader
-}
+type GenerateKubeReport = types.GenerateKubeReport
 
-type GenerateSpecReport struct {
-	Data []byte
-}
+type GenerateSpecReport = types.GenerateSpecReport
 
 type GenerateSpecOptions struct {
 	ID       string
